优化"cdist_cuda" not implemented for 'Half'
(cherry picked from commit 0cd4032b05de3d33761c5e664348f395cec36f74)

+++ src/extraction/init_center.py
@@ -6,7 +6,7 @@ import numpy as np
 def kmeans_init(features_data):
     print("In the process of initializing the center")
 
-    data = torch.tensor(features_data, device='cuda')
+    data = torch.tensor(features_data, device='cuda', dtype=torch.float32)
 
     n = len(data)
     sqrt_n = int(n ** 0.5)
